Complete overhaul of TensorAlgebra.hpp
